Implement command execution validation: add functions to extract commands from xargs and find, and validate against allowlist and denylist
