+++ src/command-validator.ts
@@ -1,5 +1,14 @@
 import { getConfig } from './config/config-loader.js';
-import { AllowCommand, DenyCommand, ShellCommandConfig } from './config/shell-command-config.js';
+import { AllowCommand, DenyCommand } from './config/shell-command-config.js';
+import {
+  findCommandInAllowlist,
+  getDenyCommandName,
+  getDenyCommandMessage,
+} from './utils/command-utils.js';
+import {
+  COMMANDS_THAT_EXECUTE_OTHER_COMMANDS,
+  validateCommandExecCommand,
+} from './command-exec-validator.js';
 
 // シェル演算子の正規表現パターン
 const SHELL_OPERATORS_REGEX = /([;|&]|&&|\|\||\(|\)|\{|\})/g;
@@ -27,14 +36,7 @@ export function checkForOutputRedirection(commandString: string): string | null 
   return null;
 }
 
-export function getCommandName(
-  commandStr: string | { command: string; subCommands?: string[] }
-): string {
-  if (typeof commandStr === 'string') {
-    return commandStr;
-  }
-  return commandStr.command;
-}
+// getCommandName function moved to utils/command-utils.js
 
 export type ValidationResult = {
   isValid: boolean;
@@ -48,137 +50,11 @@ export type ValidationResult = {
   };
 };
 
-/**
- * コマンドが許可リストに含まれているか確認する関数
- * @param commandName 確認するコマンド名
- * @param allowCommands 許可コマンドリスト
- * @returns マッチした許可コマンド、見つからない場合はnull
- */
-export function findCommandInAllowlist(
-  commandName: string,
-  allowCommands: AllowCommand[]
-): AllowCommand | null {
-  const matchedCommand = allowCommands.find((cmd) => {
-    const cmdName = getCommandName(cmd);
-    return cmdName === commandName;
-  });
+// findCommandInAllowlist function moved to utils/command-utils.js
 
-  return matchedCommand || null;
-}
+// getDenyCommandName and getDenyCommandMessage moved to appropriate utility files
 
-/**
- * DenyCommandからコマンド名を抽出
- */
-function getDenyCommandName(denyCmd: DenyCommand): string {
-  return typeof denyCmd === 'string' ? denyCmd : denyCmd.command;
-}
-
-/**
- * DenyCommandからエラーメッセージを取得する関数
- */
-function getDenyCommandMessage(denyCommand: DenyCommand, config: ShellCommandConfig): string {
-  if (typeof denyCommand === 'object' && denyCommand.message) {
-    return denyCommand.message;
-  }
-  return config.defaultErrorMessage;
-}
-
-/**
- * コマンド導入コマンドのリスト
- * xargsや-execオプションを持つfindなど、引数として他のコマンドを実行するもの
- */
-const COMMANDS_THAT_EXECUTE_OTHER_COMMANDS = ['xargs', 'find'];
-
-/**
- * xargsコマンドから実行されるコマンドを抽出する関数
- * @param command xargsを含むコマンド文字列
- * @returns 抽出されたコマンド名（見つからない場合は空文字列）
- */
-export function extractCommandFromXargs(command: string): string {
-  // xargsの後の最初の引数がコマンドとみなされる
-  const parts = command.trim().split(/\s+/);
-  const xargsIndex = parts.findIndex((part) => part === 'xargs');
-
-  if (xargsIndex >= 0 && xargsIndex + 1 < parts.length) {
-    return parts[xargsIndex + 1];
-  }
-
-  return '';
-}
-
-/**
- * find -exec/-execdir オプションから実行されるコマンドを抽出する関数
- * @param command findコマンドを含む文字列
- * @returns 抽出されたコマンド名（見つからない場合は空文字列）
- */
-export function extractCommandFromFindExec(command: string): string {
-  // -exec または -execdir オプションを検索
-  const execPattern = /\s+-exec(?:dir)?\s+(\S+)/;
-  const match = command.match(execPattern);
-
-  if (match && match[1]) {
-    return match[1];
-  }
-
-  return '';
-}
-
-/**
- * コマンド実行コマンド（xargs, find -execなど）の判定とコマンドの抽出
- * @param baseCommand ベースコマンド
- * @param command コマンド全体
- * @param config 設定オブジェクト
- * @param result 現在の結果オブジェクト
- * @returns 検証結果、問題なければnull、問題あればValidationResult
- */
-function validateCommandExecCommand(
-  baseCommand: string,
-  command: string,
-  config: ShellCommandConfig,
-  result: ValidationResult
-): ValidationResult | null {
-  // 他のコマンドを実行するコマンド（xargsやfind）の場合、引数のコマンドをチェック
-  let extractedCommand = '';
-
-  if (baseCommand === 'xargs') {
-    extractedCommand = extractCommandFromXargs(command);
-  } else if (baseCommand === 'find' && command.includes('-exec')) {
-    extractedCommand = extractCommandFromFindExec(command);
-  }
-
-  if (extractedCommand) {
-    // ブラックリストチェック
-    const blacklistedCmd = config.denyCommands.find((denyCmd) => {
-      const cmdName = getDenyCommandName(denyCmd);
-      return cmdName === extractedCommand;
-    });
-
-    if (blacklistedCmd) {
-      return {
-        ...result,
-        message: getDenyCommandMessage(blacklistedCmd, config),
-        blockReason: {
-          denyCommand: blacklistedCmd,
-          location: 'validateCommandWithArgs:blacklistedCommandInExec',
-        },
-      };
-    }
-
-    // ホワイトリストチェック
-    const matchedAllowCommand = findCommandInAllowlist(extractedCommand, config.allowCommands);
-    if (!matchedAllowCommand) {
-      return {
-        ...result,
-        message: `${config.defaultErrorMessage}: ${extractedCommand} (in ${baseCommand})`,
-        blockReason: {
-          location: 'validateCommandWithArgs:commandInExecNotInAllowlist',
-        },
-      };
-    }
-  }
-
-  return null;
-}
+// Command validation functions and constants moved to command-exec-validator.js
 
 /**
  * コマンドが許可リストに登録されているか検証する関数
@@ -224,10 +100,13 @@ export function validateCommandWithArgs(command: string): ValidationResult {
   }
 
   if (COMMANDS_THAT_EXECUTE_OTHER_COMMANDS.includes(baseCommand)) {
+    // Validate commands executed by other commands (like xargs or find -exec)
     const execCommandResult = validateCommandExecCommand(baseCommand, command, config, result);
     if (execCommandResult) {
       return execCommandResult;
     }
+
+    // Note: The old for loop that checked arguments has been removed as it's now handled by validateCommandExecCommand
   }
 
   // 直接findCommandInAllowlistを使って許可リストチェックを行う
